requirements.txt
langchain-core>=0.1.27
langchain-openai>=0.0.8
langgraph>=0.0.24
openai>=1.12.0
pandas>=2.2.1
numpy>=1.26.4
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.6.3 